feat: upgrade Influnce SDK
Ship and Building creating processes have been added to the underlying data model.
This causes a class with Process output types previously having been Products and
I patched Buildings in as Products previously. This functionality is now removed,
pending redesign to properly address for the Ship and Building types in this new
scenario. So Building is not usable as a Product now.